Moved settings to the applications directory.
