+++ src/applications/sheet-settings/parts/CheckboxSetting.svelte
