+++ src/applications/sheet-settings/parts/NumberInputSetting.svelte
