+++ src/applications/sheet-settings/tabs/InfoTab.svelte
