+++ src/applications/sheet-settings/tabs/LockSettingsTab.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import { FoundryAdapter } from 'src/foundry/foundry-adapter';
-  import CheckboxSetting from 'src/sheets/settings/parts/CheckboxSetting.svelte';
+  import CheckboxSetting from 'src/applications/sheet-settings/parts/CheckboxSetting.svelte';
   import { getContext } from 'svelte';
   import type { Writable } from 'svelte/store';
   import type { CurrentSettings } from 'src/settings/settings';
