+++ src/applications/sheet-settings/tabs/VehicleSettingsTab.svelte
@@ -1,13 +1,13 @@
 <script lang="ts">
   import { FoundryAdapter } from 'src/foundry/foundry-adapter';
-  import CheckboxSetting from 'src/sheets/settings/parts/CheckboxSetting.svelte';
+  import CheckboxSetting from 'src/applications/sheet-settings/parts/CheckboxSetting.svelte';
   import { getContext } from 'svelte';
   import type { Writable } from 'svelte/store';
   import {
     SettingsProvider,
     type CurrentSettings,
   } from 'src/settings/settings';
-  import NumberInputSetting from 'src/sheets/settings/parts/NumberInputSetting.svelte';
+  import NumberInputSetting from 'src/applications/sheet-settings/parts/NumberInputSetting.svelte';
   import SelectSetting from '../parts/SelectSetting.svelte';
 
   let context = getContext<Writable<CurrentSettings>>('context');
